uvatool: use argparser for argument parsing
Verbose but it makes it easier to insert new commandline switches. I'm
considering inotify support for executing testcases each time a file is
saved and an option to choose between w3m and firefox for displaying
problem descriptions.

Would be great if I could unifiy the str/int conversions for problem ids
as well as the handling of paths - uri, abspath, relpath...